# tests/contract/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorOrValidationResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ValidationErrorResponse"
          }
        ]
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Structured errors emitted by application and cross-cutting middleware.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "erro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erro"
          },
          "error": {
            "anyOf": [
              {
                "const": "rate_limited",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "retry_after": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvolucaoSaida": {
        "additionalProperties": false,
        "properties": {
          "banca_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banca Id"
          },
          "banca_nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banca Nome"
          },
          "lucro_acumulado_centavos": {
            "title": "Lucro Acumulado Centavos",
            "type": "integer"
          },
          "lucro_periodo_centavos": {
            "title": "Lucro Periodo Centavos",
            "type": "integer"
          },
          "periodo_inicio": {
            "format": "date",
            "title": "Periodo Inicio",
            "type": "string"
          },
          "saldo_centavos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saldo Centavos"
          }
        },
        "required": [
          "periodo_inicio",
          "banca_id",
          "banca_nome",
          "lucro_periodo_centavos",
          "lucro_acumulado_centavos",
          "saldo_centavos"
        ],
        "title": "EvolucaoSaida",
        "type": "object"
      },
...
      "MetricasGraficosSaida": {
        "additionalProperties": false,
        "properties": {
          "atualizado_em": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Em"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetSaida"
            },
            "title": "Datasets",
            "type": "array"
          },
          "granularidade": {
            "enum": [
              "dia",
              "semana",
              "mes"
            ],
            "title": "Granularidade",
            "type": "string"
          },
          "idade_mv_segundos": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idade Mv Segundos"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "replica_atraso_disponivel": {
            "title": "Replica Atraso Disponivel",
            "type": "boolean"
          },
          "replica_atraso_estado": {
            "enum": [
              "disponivel",
              "primario_ou_sem_telemetria"
            ],
            "title": "Replica Atraso Estado",
            "type": "string"
          },
          "replica_atraso_segundos": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replica Atraso Segundos"
          },
          "respondido_em": {
            "format": "date-time",
            "title": "Respondido Em",
            "type": "string"
          }
        },
        "required": [
          "atualizado_em",
          "idade_mv_segundos",
          "respondido_em",
          "replica_atraso_segundos",
          "replica_atraso_disponivel",
          "replica_atraso_estado",
          "granularidade",
          "labels",
          "datasets"
        ],
        "title": "MetricasGraficosSaida",
        "type": "object"
      },
...
      "PainelSaida": {
        "additionalProperties": false,
        "properties": {
          "atualizado_em": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Em"
          },
          "evolucao": {
            "items": {
              "$ref": "#/components/schemas/EvolucaoSaida"
            },
            "title": "Evolucao",
            "type": "array"
          },
          "idade_mv_segundos": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idade Mv Segundos"
          },
          "por_casa": {
            "items": {
              "$ref": "#/components/schemas/GrupoSaida"
            },
            "title": "Por Casa",
            "type": "array"
          },
          "por_mercado": {
            "items": {
              "$ref": "#/components/schemas/GrupoSaida"
            },
            "title": "Por Mercado",
            "type": "array"
          },
          "por_tipster": {
            "items": {
              "$ref": "#/components/schemas/GrupoSaida"
            },
            "title": "Por Tipster",
            "type": "array"
          },
          "replica_atraso_disponivel": {
            "title": "Replica Atraso Disponivel",
            "type": "boolean"
          },
          "replica_atraso_estado": {
            "enum": [
              "disponivel",
              "primario_ou_sem_telemetria"
            ],
            "title": "Replica Atraso Estado",
            "type": "string"
          },
          "replica_atraso_segundos": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replica Atraso Segundos"
          },
          "respondido_em": {
            "format": "date-time",
            "title": "Respondido Em",
            "type": "string"
          },
          "resumo": {
            "$ref": "#/components/schemas/MetricasSaida"
          },
          "saldo": {
            "$ref": "#/components/schemas/bancaemdia__api__v1__painel__SaldoSaida"
          }
        },
        "required": [
          "atualizado_em",
          "idade_mv_segundos",
          "respondido_em",
          "replica_atraso_segundos",
          "replica_atraso_disponivel",
          "replica_atraso_estado",
          "resumo",
          "saldo",
          "por_casa",
          "por_tipster",
          "por_mercado",
          "evolucao"
        ],
        "title": "PainelSaida",
        "type": "object"
      },
...
      "PeriodoPainel": {
        "enum": [
          "7d",
          "30d",
          "90d",
          "1y",
          "all"
        ],
        "title": "PeriodoPainel",
        "type": "string"
      },
...
      "ValidationErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueResponse"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CollectionToken": {
        "description": "Token opaco da extensão; armazenado no servidor somente como HMAC.",
        "in": "header",
        "name": "X-Coleta-Token",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API for authenticated betting, cash-flow, review, upload, collection, and dashboard workflows. The checked-in OpenAPI document is the compatibility contract.",
    "title": "Bancaemdia API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/painel": {
      "get": {
        "description": "Lê agregados materializados e informa separadamente frescor dos dados e atraso da réplica.",
        "operationId": "consultar_painel_api_v1_painel_get",
        "parameters": [
          {
            "description": "Janela civil de agregação no fuso America/Sao_Paulo.",
            "in": "query",
            "name": "periodo",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PeriodoPainel",
              "default": "30d",
              "description": "Janela civil no fuso America/Sao_Paulo"
            }
          },
          {
            "description": "Identificador canônico da casa usada como filtro.",
            "in": "query",
            "name": "casa_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Casa Id"
            }
          },
          {
            "description": "Identificador canônico do tipster usado como filtro.",
            "in": "query",
            "name": "tipster_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipster Id"
            }
          },
          {
            "description": "Identificador canônico do mercado usado como filtro.",
            "in": "query",
            "name": "mercado_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mercado Id"
            }
          },
          {
            "description": "Lê no primário quando verdadeiro, sem forçar refresh das materialized views.",
            "in": "query",
            "name": "fresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "Lê as mesmas MVs no primário; não altera o horário do último refresh",
              "title": "Fresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PainelSaida"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, expired, or invalid bearer token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOrValidationResponse"
                }
              }
            },
            "description": "The request does not satisfy the published contract or a domain rule."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller exceeded an applicable request limit."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required dependency is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consultar painel financeiro"
      }
    },
    "/api/v1/painel/export": {
      "get": {
        "description": "Gera um XLSX write-only e o transmite sem manter o arquivo completo em memória.",
        "operationId": "exportar_painel_api_v1_painel_export_get",
        "parameters": [
          {
            "description": "Janela civil de agregação no fuso America/Sao_Paulo.",
            "in": "query",
            "name": "periodo",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PeriodoPainel",
              "default": "30d",
              "description": "Janela civil no fuso America/Sao_Paulo"
            }
          },
          {
            "description": "Identificador canônico da casa usada como filtro.",
            "in": "query",
            "name": "casa_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Casa Id"
            }
          },
          {
            "description": "Identificador canônico do tipster usado como filtro.",
            "in": "query",
            "name": "tipster_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipster Id"
            }
          },
          {
            "description": "Identificador canônico do mercado usado como filtro.",
            "in": "query",
            "name": "mercado_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mercado Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Streaming Excel workbook."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, expired, or invalid bearer token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOrValidationResponse"
                }
              }
            },
            "description": "The request does not satisfy the published contract or a domain rule."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller exceeded an applicable request limit."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required dependency is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Exportar painel em Excel"
      }
    },
    "/api/v1/painel/metricas": {
      "get": {
        "description": "Retorna séries temporais já agregadas para os gráficos do painel.",
        "operationId": "consultar_metricas_api_v1_painel_metricas_get",
        "parameters": [
          {
            "description": "Janela civil de agregação no fuso America/Sao_Paulo.",
            "in": "query",
            "name": "periodo",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PeriodoPainel",
              "default": "30d",
              "description": "Janela civil no fuso America/Sao_Paulo"
            }
          },
          {
            "description": "Identificador canônico da casa usada como filtro.",
            "in": "query",
            "name": "casa_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Casa Id"
            }
          },
          {
            "description": "Identificador canônico do tipster usado como filtro.",
            "in": "query",
            "name": "tipster_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipster Id"
            }
          },
          {
            "description": "Identificador canônico do mercado usado como filtro.",
            "in": "query",
            "name": "mercado_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mercado Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricasGraficosSaida"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, expired, or invalid bearer token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOrValidationResponse"
                }
              }
            },
            "description": "The request does not satisfy the published contract or a domain rule."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller exceeded an applicable request limit."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required dependency is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consultar séries para gráficos"
      }
    },
...
  }
}